Refactor fetch pokemon api and list all users
